requirements.txt
Pillow==8.4.0
django-froala-editor==3.2.2